meta/test: use random db name for sqlite tests
