+++ pkg/meta/sql_test.go
@@ -17,13 +17,13 @@
 package meta
 
 import (
-	"os"
 	"testing"
+
+	"github.com/juicedata/juicefs/pkg/utils"
 )
 
 func TestSQLClient(t *testing.T) {
-	os.Remove("test.db")
-	m, err := newSQLMeta("sqlite3", "test.db", &Config{})
+	m, err := newSQLMeta("sqlite3", utils.RandString(8)+".db", &Config{})
 	if err != nil {
 		t.Fatalf("create meta: %s", err)
 	}
@@ -50,8 +50,7 @@ func TestMySQLClient(t *testing.T) {
 }
 
 func TestStickyBitSQL(t *testing.T) {
-	os.Remove("test3.db")
-	m, err := newSQLMeta("sqlite3", "test3.db", &Config{})
+	m, err := newSQLMeta("sqlite3", utils.RandString(8)+".db", &Config{})
 	if err != nil {
 		t.Fatalf("create meta: %s", err)
 	}
@@ -59,8 +58,7 @@ func TestStickyBitSQL(t *testing.T) {
 }
 
 func TestLocksSQL(t *testing.T) {
-	os.Remove("test1.db")
-	m, err := newSQLMeta("sqlite3", "test1.db", &Config{})
+	m, err := newSQLMeta("sqlite3", utils.RandString(8)+".db", &Config{})
 	if err != nil {
 		t.Fatalf("create meta: %s", err)
 	}
@@ -68,8 +66,7 @@ func TestLocksSQL(t *testing.T) {
 }
 
 func TestConcurrentWriteSQL(t *testing.T) {
-	os.Remove("test1.db")
-	m, err := newSQLMeta("sqlite3", "test1.db", &Config{})
+	m, err := newSQLMeta("sqlite3", utils.RandString(8)+".db", &Config{})
 	if err != nil {
 		t.Fatalf("create meta: %s", err)
 	}
@@ -77,8 +74,7 @@ func TestConcurrentWriteSQL(t *testing.T) {
 }
 
 func TestCompactionSQL(t *testing.T) {
-	os.Remove("test2.db")
-	m, err := newSQLMeta("sqlite3", "test2.db", &Config{})
+	m, err := newSQLMeta("sqlite3", utils.RandString(8)+".db", &Config{})
 	if err != nil {
 		t.Fatalf("create meta: %s", err)
 	}
@@ -86,8 +82,7 @@ func TestCompactionSQL(t *testing.T) {
 }
 
 func TestTruncateAndDeleteSQL(t *testing.T) {
-	os.Remove("test.db")
-	m, err := newSQLMeta("sqlite3", "test.db", &Config{})
+	m, err := newSQLMeta("sqlite3", utils.RandString(8)+".db", &Config{})
 	if err != nil {
 		t.Fatalf("create meta: %s", err)
 	}
@@ -95,8 +90,7 @@ func TestTruncateAndDeleteSQL(t *testing.T) {
 }
 
 func TestCopyFileRangeSQL(t *testing.T) {
-	os.Remove("test.db")
-	m, err := newSQLMeta("sqlite3", "test.db", &Config{})
+	m, err := newSQLMeta("sqlite3", utils.RandString(8)+".db", &Config{})
 	if err != nil {
 		t.Fatalf("create meta: %s", err)
 	}
@@ -104,8 +98,7 @@ func TestCopyFileRangeSQL(t *testing.T) {
 }
 
 func TestCaseIncensiSQL(t *testing.T) {
-	os.Remove("test.db")
-	m, err := newSQLMeta("sqlite3", "test.db", &Config{CaseInsensi: true})
+	m, err := newSQLMeta("sqlite3", utils.RandString(8)+".db", &Config{CaseInsensi: true})
 	if err != nil {
 		t.Fatalf("create meta: %s", err)
 	}
